Merging sprint3backend to main

+++ backend/database/reviews.py
@@ -1,8 +1,11 @@
 import oracledb
 #import connect
-from database import connect
+
+#MAJOR NOTE: This import is what works for both render (where the backend is hosted) and at least for me (andrew) locally
+from database import connect #If this import doesn't work change it temporarily, but be sure to change it back before pushing because the backend won't understand "import connect" when it's on render
+
 #from users import valid_user_id
-from .users import valid_user_id
+from database.users import valid_user_id
 
 def format_review(row):
     return {
@@ -20,7 +23,7 @@ def get_new_review_id():
         print("Failed to connect to database.")
         return None
 
-    cursor.execute("SELECT MAX(REVIEW_ID) FROM REVIEWS")
+    cursor.execute("SELECT MAX(REVIEW_ID) FROM ADMIN.REVIEWS")
     result = cursor.fetchone()
 
     connect.stop_connection(connection, cursor)
@@ -39,18 +42,18 @@ def add_review(user_id, media_id, media_type, rating, review_text):
     if valid_user_id(user_id):
         try:
             if media_type == "book":
-                cursor.execute("SELECT get_book_id(:1) FROM dual", (media_id,))
+                cursor.execute("SELECT ADMIN.get_book_id(:1) FROM dual", (media_id,))
                 result = cursor.fetchone()
                 db_media_id = result[0]
             else:
-                db_media_id = media_id
+                db_media_id = int(media_id)
 
             cursor.execute(
                 """
-                INSERT INTO REVIEWS (REVIEW_ID, USER_ID, MEDIA_ID, MEDIA_TYPE, RATING, REVIEW_TEXT)
+                INSERT INTO ADMIN.REVIEWS (REVIEW_ID, USER_ID, MEDIA_ID, MEDIA_TYPE, RATING, REVIEW_TEXT)
                 VALUES (:1, :2, :3, :4, :5, :6)
                 """,
-                (review_id, user_id, media_id, media_type, rating, review_text)
+                (review_id, user_id, db_media_id, media_type, rating, review_text)
             )
             connection.commit()
             print("Review added successfully.")
@@ -85,7 +88,7 @@ def delete_review(review_id):
     try:
         cursor.execute(
             """
-            DELETE FROM REVIEWS WHERE REVIEW_ID = :1
+            DELETE FROM ADMIN.REVIEWS WHERE REVIEW_ID = :1
             """,
             (review_id,)
         )
@@ -107,7 +110,7 @@ def delete_review(review_id):
 
 def print_reviews():
     connection, cursor = connect.start_connection()
-    cursor.execute("SELECT * FROM REVIEWS")
+    cursor.execute("SELECT * FROM ADMIN.REVIEWS")
     row = cursor.fetchall()
     if row:
         for row in row:
@@ -122,7 +125,7 @@ def get_all_reviews():
         return None
 
     try:
-        cursor.execute("SELECT * FROM REVIEWS")
+        cursor.execute("SELECT * FROM ADMIN.REVIEWS")
         rows = cursor.fetchall()
         reviews = []
         for row in rows:
@@ -152,7 +155,7 @@ def get_reviews_by_media_type(media_type):
         print("Failed to connect to database.")
         return None
     try:
-        cursor.execute("SELECT * FROM REVIEWS WHERE MEDIA_TYPE = :1", (media_type,))
+        cursor.execute("SELECT * FROM ADMIN.REVIEWS WHERE MEDIA_TYPE = :1", (media_type,))
         rows = cursor.fetchall()
         return rows
     except oracledb.Error as e:
@@ -170,12 +173,12 @@ def get_reviews_by_media_id_and_type(media_id, media_type):
         return None
     try:
         if media_type == "book":
-            cursor.execute("SELECT get_book_id(:1) FROM REVIEWS", (media_id,))
+            cursor.execute("SELECT ADMIN.get_book_id(:1) FROM dual", (media_id,))
             result = cursor.fetchone()
             db_media_id = result[0]
         else:
             db_media_id = media_id
-        cursor.execute("SELECT * FROM REVIEWS WHERE MEDIA_ID = :1 AND MEDIA_TYPE = :2", (db_media_id, media_type))
+        cursor.execute("SELECT * FROM ADMIN.REVIEWS WHERE MEDIA_ID = :1 AND MEDIA_TYPE = :2", (db_media_id, media_type))
         rows = cursor.fetchall()
 
         reviews = []
@@ -197,7 +200,7 @@ def get_reviews_by_user_id(user_id):
         print("Failed to connect to database.")
         return None
     try:
-        cursor.execute("SELECT * FROM REVIEWS WHERE USER_ID = :1", (user_id,))
+        cursor.execute("SELECT * FROM ADMIN.REVIEWS WHERE USER_ID = :1", (user_id,))
         rows = cursor.fetchall()
 
         reviews = []
@@ -222,7 +225,7 @@ def get_reviews_by_user_id_and_media_type(user_id, media_type):
         print("Failed to connect to database.")
         return None
     try:
-        cursor.execute("SELECT * FROM REVIEWS WHERE USER_ID = :1 AND MEDIA_TYPE = :2", (user_id,media_type))
+        cursor.execute("SELECT * FROM ADMIN.REVIEWS WHERE USER_ID = :1 AND MEDIA_TYPE = :2", (user_id,media_type))
         rows = cursor.fetchall()
 
         #Format reviews
@@ -247,12 +250,12 @@ def get_reviews_by_user_id_and_media_id_and_media_type(user_id, media_id, media_
         return None
     try:
         if media_type == "book":
-            cursor.execute("SELECT get_book_id(:1) FROM REVIEWS", (media_id,))
+            cursor.execute("SELECT ADMIN.get_book_id(:1) FROM REVIEWS", (media_id,))
             result = cursor.fetchone()
             db_media_id = result[0]
         else:
             db_media_id = media_id
-        cursor.execute("SELECT * FROM REVIEWS WHERE USER_ID = :1 AND MEDIA_ID = :2 AND MEDIA_TYPE = :3", (user_id,media_id, media_type))
+        cursor.execute("SELECT * FROM ADMIN.REVIEWS WHERE USER_ID = :1 AND MEDIA_ID = :2 AND MEDIA_TYPE = :3", (user_id,media_id, media_type))
         rows = cursor.fetchall()
 
         reviews = []
@@ -277,7 +280,7 @@ def edit_review(review_id, review_text):
     try:
         cursor.execute(
             """
-            UPDATE REVIEWS
+            UPDATE ADMIN.REVIEWS
             SET REVIEW_TEXT = :1
             WHERE REVIEW_ID = :2
             """,
@@ -299,12 +302,3 @@ def edit_review(review_id, review_text):
 
     finally:
         connect.stop_connection(connection, cursor)
-
-
-#print_reviews()
-#add_review(4,1,"a",5,"")
-#print_reviews()
-#delete_review(1)
-#print_reviews()
-#edit_review(1, "best movie ever omg!")
-#print_reviews()
